server/pyproject.toml
[project]
name = "cells-interlinked"
version = "0.1.0"
description = "A Voight-Kampff test for language models"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.4",
    "transformers>=4.51",
    "accelerate>=0.30",
    "safetensors>=0.4",
    "huggingface-hub>=0.24",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sse-starlette>=2.1",
    "aiosqlite>=0.20",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    "numpy>=1.26",
]

[dependency-groups]
dev = [
    "ruff>=0.6",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cells_interlinked"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "RUF"]
